RUN-4373 Window groups to Typescript (#521)
* converted to typescript

* fixed tslint comments

* added super call

* code review items

* code review
