+++ src/browser/api/window.js
@@ -28,7 +28,7 @@ import { cachedFetch } from '../cached_resource_fetcher';
 let log = require('../log');
 import ofEvents from '../of_events';
 import SubscriptionManager from '../subscription_manager';
-let WindowGroups = require('../window_groups.js');
+import WindowGroups from '../window_groups';
 import { addConsoleMessageToRVMMessageQueue } from '../rvm/utils';
 import { validateNavigation, navigationValidator } from '../navigation_validation';
 import { toSafeInt } from '../../common/safe_int';
